add simulation delay and rename the component
