+++ src/App.tsx
@@ -1,6 +1,6 @@
 import FetchIPAddress from './components/DNS/fetchIPAddress';
 import FetchData from './components/Fetch_API/fetchCall';
-import UrlViewer from './components/URL/urlParts';
+import UrlViewer from './components/URL | Async/urlParts';
 
 export default function App() {
     return (
